requirements.txt
langchain
langchain_experimental
langchain_openai
python-dotenv
pandas
tabulate
flask
fastapi[standard]